requirements.txt
Django==1.8
django-filter==0.9.2
djangorestframework>=3.0
django-grappelli>=2.6.5
numpy>=1.6
astropy(>=0.4)
astroquery>=2.0
Pillow==3.3.1
MySQL-python==1.2.5